fix(CORE/Creatures): Preserve rounded creature base armor
GenerateArmor applies the creature's fractional armor modifier and rounds the result. Calling it through a float initialization discarded that rounding comment and obscured the integer value being assigned to the armor modifier. Keep the returned uint32 and convert once when storing the BASE_VALUE.

+++ src/server/game/Entities/Creature/Creature.cpp
@@ -547,8 +547,10 @@ bool Creature::UpdateEntry(uint32 Entry, const CreatureData* data, bool changele
 
     SetMeleeDamageSchool(SpellSchools(cInfo->dmgschool));
     CreatureBaseStats const* stats = sObjectMgr->GetCreatureBaseStats(getLevel(), cInfo->unit_class);
-    float armor = (float)stats->GenerateArmor(cInfo); // TODO: Why is this treated as uint32 when it's a float?
-    SetModifierValue(UNIT_MOD_ARMOR,             BASE_VALUE, armor);
+    // GenerateArmor applies a fractional armor modifier and rounds the
+    // result, so keep the integer result instead of truncating it via float.
+    uint32 armor = stats->GenerateArmor(cInfo);
+    SetModifierValue(UNIT_MOD_ARMOR,             BASE_VALUE, float(armor));
     SetModifierValue(UNIT_MOD_RESISTANCE_HOLY,   BASE_VALUE, float(cInfo->resistance[SPELL_SCHOOL_HOLY]));
     SetModifierValue(UNIT_MOD_RESISTANCE_FIRE,   BASE_VALUE, float(cInfo->resistance[SPELL_SCHOOL_FIRE]));
     SetModifierValue(UNIT_MOD_RESISTANCE_NATURE, BASE_VALUE, float(cInfo->resistance[SPELL_SCHOOL_NATURE]));
